Rename: core 폴더 -> api 폴더
